chore: do not use tm modified pkg files
Instead of moving modified traefik/traefik function into a separate file the origin now stays the same. Also the func names will not change to the official. There are to many changes (also for next features) to use separate approach.
