+++ .golangci.yml
@@ -70,7 +70,4 @@ issues:
         - godot
 
   exclude-files:
-    - pkg/github.com/traefik/traefik/pkg/config/dynamic/middlewares.go
-    - pkg/github.com/traefik/traefik/pkg/ip/checker.go
-    - pkg/github.com/traefik/traefik/pkg/ip/strategy.go
-
+    - pkg/github.com/traefik/traefik/
